refactor: extract board beads module
Also completes lib/sse.mjs by moving broadcastTasks in from server.mjs
(deferred from the earlier sse extraction step since it depends on
getTasks, which only now exists as a standalone module). data-readers.mjs's
getTasks import is repointed from the temporary server.mjs circular import
to lib/beads.mjs; the readVerdicts/readSecStats circular import to
server.mjs remains until the verdicts.mjs extraction step.
