+++ packages/board/lib/sse.mjs
@@ -1,4 +1,5 @@
 import { sseClients } from './state.mjs';
+import { getTasks } from './beads.mjs';
 
 function broadcast(event, data) {
   const msg = `event: ${event}\ndata: ${JSON.stringify(data)}\n\n`;
@@ -7,4 +8,13 @@ function broadcast(event, data) {
   }
 }
 
-export { broadcast };
+function broadcastTasks(cwd) {
+  const msg = `event: tasks\ndata: ${JSON.stringify(getTasks(cwd))}\n\n`;
+  for (const res of sseClients) {
+    if (res._gctoCwd === cwd) {
+      try { res.write(msg); } catch { sseClients.delete(res); }
+    }
+  }
+}
+
+export { broadcast, broadcastTasks };
